Improve scope handling. Allow anonymous functions. Work in progress.
Frames are now managed in a tree fashion rather than a list.
Each time a new level of namespace is created, a new branch of frames
is forked. For example, between the global namespace and the local one,
or between a local namespace and a nested function. Access to variables
in upper namespaces is performed through indirections in the frames
ancestorship. Note that a new frame is created a each function call,
but not necessarily a new branch. When a new frame is created, it
inherits its ancestor from the caller frame, so only one indirection
is necessary to access a global var, even in a deep recursive call.

For each named variable, the compiler stores the frame index and the
frame level, i.e. the number of indirections in the frame tree.