fix: propagate correct type for CompositeLitExpr
At CFG, in pre-order processing, determine the correct type of
CompositeLitExpr from its first child (if it's a type) or from
the ancestor node. Make sure The type is propagated to children
so the algorithm works recursively.

Fix also the isType() method to handle case of imported types,
either from source or binary packages.

Add all previous tests as in the review