feat: support relative import paths "../xxx" and "./xxx"
Relative import paths are now detected in importSrcFile().
The initial base dir is the dir of input file or current working
directory. The logic of relative import path is completely
separate of the one for absolute import path, and could be
easily disabled, if we decide so.

Additionally, error reporting has been slightly improved in
importSrcFile.

Fix #145