fix: resolve binary methods in interface wrappers
In interface wrapper, if interpreted method is not found for the given
interface, search for binary method, which can be embedded in struct
fields.

In function wrapper, handle setting the receiver in the case the
expected receiver is a pointer, and the passed one is not.

Also fix appendSlice which was not setting the result correctly.