feature: set default GOPATH for command line and allow self interpreting
To be on par with "go run", the default GOPATH is applied to yaegi
command line (note that interp default behaviour is unchanged).

Flag state is explicitely initialized prior to start the REPL, to
allow recursion.

Make sure that stdlib symbols define themselves to allow recursion.

It is now possible to perform the following:

```console
$ go build cmd/yaegi/yaegi.go
$ ./yaegi cmd/yaegi/yaegi.go cmd/yaegi/yaegi.go cmd/yaegi/yaegi.go
>
```