fix: do not skip AST errors when not in REPL mode
The trick to use the Go AST parser and implement the REPL mode,
which requires incremental, line by line parsing, is to wrap
statements into package or functions.

It is now performed only when explicetly in REPL mode, allowing to catch
much more AST errors in batch mode.

Problems of stuck REPL in case of wrong input is outside the
scope of this change, and must be addressed separately, by
more detailed AST error checking in REPL mode.

Fix #238