fix: global variable init from selector expression
In nodeType(), in presence of a selector expression, the logic was broken
for complex expressions. We now first compute the type of the left part
of the selector, then perform a lookup operation on it.

We now store the package path in type rather than symbol, to allow the
above on srcPkgT and binPkgT.

Fix #359