fix: correct automatic type conversion for untyped constants
For the following operators: <<, >>, <<=, >>=, untyped constant
values are accepted if they are convertible to int and unsigned.

For example, `1 << 1.0` and `1.0 << 1` are valid by `1 << 1.1`
is not.

This change ensures the required properties are checked for untyped
constants and that result is properly converted and propagated.
Unit tests are updated to match the behaviour of the official go compiler.

Note: in the official go compiler, there may be an inconsistency in
the way untyped constants are processed, as the above behaviour
applies only for shift operations, not other bits operations. For
example, `1.0 & 1` in forbidden.

Fix #278