feature: add Stop() and EvalTimeout()
This is an alternate solution to `context.Context`, to implement
eval with timeout and cancellation, as per #386

This solution uses a `runid` counter added both to the execution
frames and the global `Interpreter` struct. When Eval starts, the
current `runid` from `Interpreter` is propagated to the frames
(a new frame is created at each function call, including closures
and goroutines; a frame inherits the `runid` from its ancestor).
in frames, the runid is written once at frame init, and never modified.
The interpreter runid can be modified to trigger the stop condition
(see Stop()).

Eval stop is triggered when the current frame runid is different
from the interpreter runid. This stop condition is evaluated in the
execution loop of `runCfg` at each instruction step. From my measures,
this integer equality test is 20 times faster than a select statement
on closed channel.

An interpreter session consists of multiple evals. Eval can spawn
a set of background goroutines. With this algorithm, all active
goroutines will be stopped, no matter which previous eval started
them. After a stop, a new Eval can happen.

To be more complete, a mechanism to unblock channels read/write, select
and locks must be added as well, this is planned in a further change.
This can be similar to a closed channel broadcast (as in context.Context)
except that the stop channel will be probably defined once in Interpreter
struct, and be used only for blocking calls, not execution loop. PR #387
proposed at some point an implementation of such interruptible blocking
calls.

The unit tests have been copied from PR #387.

the REPL has been modified to allow Eval interrupt by CTRL-C using
the above mechanism.

This is experimental at this point, and for discussion and comparison
with #387.