feature: keep both blocking and cancellable channel I/O
An interpreter option, `cancelChan` triggers the generation of
slow but cancellable channel read / write when true, or fast blocking
otherwise.

`cancelChan` is set to `true` in REPL mode, `false` otherwise. Note
that the external behaviour remains the same, as `evalWithContext`
will always unblock. The difference relies in the capacity to release
gooroutines hanged on channels (at least the interpreted ones).