feature: by default evalWithContext activates cancellable chan op
This can be disabled by setting YAEGI_FAST_CHAN environment variable.
Note that if only Eval() is used, and not EvalWithContext(), the
default behaviour is to use the fast blocking version of chan
operations (as in non-interactive scripts).

As soon as EvalWithContext() is used, even only once (for example in
interactive or embedded mode), the interpreter switches to cancellable
chan operations, unless YAEGI_FAST_CHAN is set.

In summary, default is fast blocking for non-interactive/batch, or
cancellable for interactive/embedded.