fix: handle use of functions in struct fields
Functions in struct fields are wrapped in binary functions, in
order to allow interchange between runtime and interpreter context.
But in this process, the original type of function returned objects
is lost (replaced by anonymous reflect types), making impossible to
perform method resolution (based on type names).

The solution to this problem is to "unwrap" struct field function
types, to preserve the necessary information, and to keep along
the original type reference.

Fixes #525