fix: better handling of recursive types and forward declarations
As several GTA passes are required to handle forward declarations
and recursive types, make sure to update the type recursive status
in the type symbol, using scope to retrieve the symbol. The scope
was not always recorded at type creation. Fix that for dynamic
type creation which are potentially recursive.

Update refType() to return and interface{} in place of struct to
represent a recursive type using reflect.

Fixes #570.