fix: assign composite literal by reference
When using reflect.Set() to assign a struct object to a variable, all
data is duplicated, but the underlying struct address in destination is
not affected, which is different of allocating a new struct.  The right
behaviour is to set directly the reflect.Value in the frame, without
using Set.

Fixes #579.