interp: global scope redeclaration detection
Part 1, some cases are still not handled, such as:

- if the nth (n != 1) declaration is a typeSpec, it does not seem easy
to handle it, without breaking (out of order) type declarations that
build upon another type defined in the same file.

- methods haven't been investigated yet. No idea on complexity.

- non stdlib imports haven't been done yet. I do not expect them to be
harder than stdlib ones.

One notable change of this PR is how the symbols for imports are now
keyed by identifier+suffix, where suffix is the current go source file
base name. Whereas previously the key was only the identifier, and since
imports symbols are in the global scope, it means two imports of the
same package name, each of them in a different source file of the same
package would overwrite each other in the map of symbols. Out of luck,
this was not causing any visible problem, but it became apparent when
dealing with redeclaration.

Also, most of the errors returned are not as precise as the ones the go
std compiler would give, i.e. they do not always specify where the
previous declaration was.

Fixes #662