interp: refactor doComposite cases
Also, there is an optimization that used to take place when the RHS of
an assignment is a struct literal. That optimization avoids some
allocation by "instructing" that whatever we want to write for the frame
of the RHS gets directly written to the frame of the LHS.
But it just so happens that we're doing this incorrently when the LHS is
a struct field index (or probably for any other kind of getIndex AST
action). Therefore this PR also disables the optimization when the LHS
is a getIndex.

Fixes #851
Fixes #854