interp: fix a regression in type checks
When type checking assignments or conversions, we may have to check
if a constant value, known at compile time, is representable in a given type.
It involves retrieving the value from the node, which is not always
populated in type object. The assignableTo type method now requires the node
value to be passed explicitely, not relying on how the type was
obtained.

Fixes #1238.