interp: virtualize environment in restricted mode
In restricted mode, replace environment related symbols in
stdlib os package by a version which operate on a private copy
per interpreter context.

It allows to have concurrent interpreters in the same process
operating each in their own environment without affecting each
other or the host.

If YAEGI_UNRESTRICTED=1, this behaviour is disabled, and the
default symbols from stdlib are used.

Note also that modification is done for syscall package, as it
should be not used in restricted mode.