feat: support go1.18 standard library
Warappers for go1.16 are removed, wrappers for go1.18 are added.
TestInterpConsistencyBuild() is removed due to too many differences
between go1.17 and go1.18 behaviour about error messages, and the fact
that it is mostly redundant with TestFile(), but slower and more complex.

Note that there is no change in the interpreter, in particular
no support for generics yet, as the parser would be incompatible
with go1.17. This will be provided when go1.19 is released.