interp: fix type switch on arbitrary expressions
If the value on which to type-switch was already set (i.e. a variable),
there was no problem. But if it had to be obtained through a complex
expression (func call, array index, etc...), then the code to retrieve
the value prior type-switch was not scheduled. This is now fixed.

The issue it fixes is nasty because the behavior was silently changed,
leading to further unrelated issues or panics.

Fixes #1444.