interp: improve handling of methods defined on interfaces
For methods defined on interfaces (vs concrete methods), the resolution
of the method is necessarily delayed at the run time and can not be
completed at compile time.

The selectorExpr processing has been changed to correctly identify
calls on interface methods which were confused as fields rather
than methods (due to the fact that in a interface definition, methods
are fields of the interface).

Then at runtime, method lookup has been fixed to correctly recurse in
nested valueInterface wrappers and to find embedded interface fields
in case of struct or pointer to structs.

Fixes #1515.